Allow to build different spannable objects when ticks have different sizes
In unicode languages is possible that the same time format produce tick texts
with different sizes. Create one DynamicSpannableString for every tick text size,
so we don't have to recrete objects every time. Actually, it got a FC because
we are trying to update the current DynamicSpannableString ref, which its
structures are prepared for other size